Implementación de ViewModels, estados y nueva pantalla de precios
- Se implementan los ViewModels `TrendsViewModel`, `ProfileViewModel`, `PricesViewModel`, `DetailViewModel`, `ConfigurationViewModel` y `CommentsViewModel` para gestionar la lógica de negocio y el estado de la UI.
- Se crean clases de estado (`State`) para las pantallas de Profile, Trends, Prices, Detail, Configuration y Comments.
- Se añade la nueva pantalla `Prices` y sus componentes relacionados para mostrar comparativas de precios de productos.
- Se actualiza `AppNavigation` para incluir la ruta de "prices" y mejorar la navegación entre detalles y comentarios.
- Se refactoriza la pantalla de `Profile` para consumir datos dinámicos desde su ViewModel en lugar de proveedores estáticos.
- Se añade `PricedItemsProvider` y recursos de imagen para cadenas de supermercados (Ara, Carulla, D1, Éxito, Olímpica).
- Se mejoran componentes de UI en `Detail` y `Configuration` para soportar interacciones y estados (búsqueda, toggles, navegación).
- Ajustes menores en descripciones de accesibilidad y diseño de componentes comunes.

+++ app/src/main/java/com/example/buy_it/data/datasource/impl/UserRetrofitDatasourceImplementation.kt
@@ -8,7 +8,7 @@ import javax.inject.Inject
 
 class UserRetrofitDatasourceImplementation @Inject constructor(private val service: UserRetrofitService): UserRemoteDatasource {
     override suspend fun getUserById(id: String): UserProfileDTO {
-        return service.getUserById(id)
+        return service.getUserById(id.toInt())
     }
 
     override suspend fun getUserReviews(id: String): List<ReviewDTO> {
